Other refactors.
