+++ app/src/main/java/com/gb/stopwatch/domain/helpers/StopwatchStateCalculator.kt
@@ -1,4 +1,4 @@
-package com.gb.stopwatch.domain
+package com.gb.stopwatch.domain.helpers
 
 import com.gb.stopwatch.domain.data.StopwatchState
