+++ app/src/main/java/com/gb/stopwatch/domain/helpers/TimestampProvider.kt
@@ -1,4 +1,4 @@
-package com.gb.stopwatch.domain
+package com.gb.stopwatch.domain.helpers
 
 interface TimestampProvider {
     fun getMilliseconds(): Long
